@@ -2,7 +2,9 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
+	  <ImplicitUsings>true</ImplicitUsings>
+	  <Nullable>enable</Nullable>
   </PropertyGroup>
 
 </Project>
