@@ -2,7 +2,8 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+	  <TargetFramework>net7.0</TargetFramework>
+	  <ImplicitUsings>true</ImplicitUsings>
   </PropertyGroup>
 
 </Project>
