@@ -1,49 +1,52 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection.Metadata.Ecma335;
+﻿namespace Methods;
 
-namespace Methods
-{
-	/*
-    #############################################################################
-    ### Metody (Funkce) aneb procedurální programování
-    #############################################################################
+/*
+#############################################################################
+### Metody (Funkce) aneb procedurální programování
+#############################################################################
         
-    Druhy metod:
-    - Statické (Voláme skrze název třídy.)
-    - Nestatické (Potřebují instanci třídy a voláme skrze název instance třídy.)
+Druhy metod:
+- Statické (Voláme skrze název třídy.)
+- Nestatické (Potřebují instanci třídy a voláme skrze název instance třídy.)
 
-    Použití metody vyžaduje:
-    - Definici funkce
-    - Zavolání funkce
+Použití metody vyžaduje:
+- Definici funkce
+- Zavolání funkce
 
-    Definice:
-    <modifikátory přístupu> <návratový typ> <Jméno> (<Seznam parametrů>) 
-    {
-       Tělo metody
-       Práce s parametry
-       Příprava návratové hodnoty
-    }
+Definice:
+<modifikátory přístupu> <návratový typ> <Jméno> (<Seznam parametrů>) 
+{
+    Tělo metody
+    Práce s parametry
+    Příprava návratové hodnoty
+}
 
     Seznam parametrů při definici:
-    ([parametr 1]: <modifikátor přístupu> <datový typ> <název hodnoty>, [parametr 2]: <modifikátor přístupu> <datový typ> <název hodnoty>)
+([parametr 1]: <modifikátor přístupu> <datový typ> <název hodnoty>, [parametr 2]: <modifikátor přístupu> <datový typ> <název hodnoty>)
 
 
-    Seznam parametrů při volání:
-    [parametr 1]: <(nepovinné) název definice parametru>: <modifikátor přístupu> <název hodnoty>, [parametr 2]: <(nepovinné) název definice parametru>: <modifikátor přístupu> <datový typ> <název hodnoty>
+Seznam parametrů při volání:
+[parametr 1]: <(nepovinné) název definice parametru>: <modifikátor přístupu> <název hodnoty>, [parametr 2]: <(nepovinné) název definice parametru>: <modifikátor přístupu> <datový typ> <název hodnoty>
 
-    public Int32 Sum(firstOperand: ref a, secondOperand: ref b) {...}
-    #############################################################################
-    */
-	class Methods
+public Int32 Sum(firstOperand: ref a, secondOperand: ref b) {...}
+#############################################################################
+*/
+class Methods
+{
+	// Povinná metoda každého C# programu.
+	static void Main(string[] args)
 	{
-		// Povinná metoda každého C# programu.
-		static void Main(string[] args)
-		{
-			
-            // Lokální proměnné
+		Console.WriteLine(SolutionNew.Factorial(1));
+		Console.WriteLine(SolutionNew.Factorial(2));
+		Console.WriteLine(SolutionNew.Factorial(3));
 
-            int a = 100;
+		SolutionNew.Fibbonaci(1);
+		SolutionNew.Fibbonaci(2);
+		SolutionNew.Fibbonaci(3);
+		SolutionNew.Fibbonaci(10);
+        // Lokální proměnné
+
+        int a = 100;
             int b = 200;
             var methods = new Methods();
             int ret = methods.FindMax(a, b);
@@ -105,133 +108,134 @@ namespace Methods
             Console.WriteLine($"Sum of {a} and {b} is {methods.Sum(in a, in b)}");
             Console.WriteLine($"Sum of {a} and {b} is {methods.Sum(a, b)}"); // U volání není nutné použít in, to se doplní samo.
 
-			methods.PrintNumber(10.12345); // Automaticky se použije 3 desetinná místa, protože to je výchozí hodnota 2. parametru.
-			methods.PrintNumber(10.12345, 2);
+		methods.PrintNumber(10.12345); // Automaticky se použije 3 desetinná místa, protože to je výchozí hodnota 2. parametru.
+		methods.PrintNumber(10.12345, 2);
 
 
-			Solution.Fibonacci(1000);
-			Solution.Fibonacci(1000, 20);
+		Solution.Fibonacci(1000);
+		Solution.Fibonacci(1000, 20);
 
-			Console.ReadLine();
-		}
+		Console.ReadLine();
+	}
 
-		// Jednoduchá veřejná funkce
-		public Int32 FindMax(Int32 num1, Int32 num2)
+	// Jednoduchá veřejná funkce
+	public Int32 FindMax(Int32 num1, Int32 num2)
+	{
+		return num1 > num2 ? num1 : num2;
+	}
+
+    // Jednoduchá veřejná funkce (komplaktní
+    public Int32 FindMaxCopmact(Int32 num1, Int32 num2) => num1 > num2 ? num1 : num2;
+
+    // Jednoduchá statická veřejná funkce
+    public static Int32 FindMaxStatic(Int32 num1, Int32 num2)
+	{
+		return num1 > num2 ? num1 : num2;
+	}
+
+	// Rekurzivní funkce
+	public Int32 Factorial(Int32 num)
+	{
+		// 0 a 1.
+		if (num < 2)
 		{
-			return num1 > num2 ? num1 : num2;
+			return 1;
 		}
 
-		// Jednoduchá statická veřejná funkce
-		public static Int32 FindMaxStatic(Int32 num1, Int32 num2)
+		return Factorial(num - 1) * num;
+		// Hrozná prasárna pro zásobník volání.
+	}
+
+	// Předávání parametrů hodnotou
+	// Dochází dvakrát k boxingu (zabalení), konkrétně
+	// hodnoty x a y.
+	public void Swap(Int32 x, Int32 y)
+	{
+		var temp = x;
+		x = y;    /* put y into x */
+		y = temp; /* put temp into y */
+		// return; // nepovinné a je odebráno při automatickém formátování kódu.
+	}
+
+	// Přetížená funkce prohození s předáním referencí
+	public void Swap(ref Int32 x, ref Int32 y)
+	{
+		var temp = x;
+		x = y;    /* put y into x */
+		y = temp; /* put temp into y */
+	}
+
+	// Parametry definované s ref musí být při volání funkce
+	// definovány taktéž jako ref, tj. jako reference.
+	public void SwapRef(ref Int32 x, ref Int32 y)
+	{
+		var temp = x;
+		x = y;    /* put y into x */
+		y = temp; /* put temp into y */
+	}
+
+	// Několika násobný výstup z funkce
+	// Klíčové slovo out říká kompilátoru,
+	// že proměnná bude inicializována později.
+	public Int32 GetSomeValue(out Int32 x)
+	{
+		x = 5;
+		return x;
+	}
+
+	// Definice swap funkce (prohození) pomocí ref a out parametrů.
+	public void SwapOut(ref Int32 a, ref Int32 b, out Int32 x, out Int32 y)
+	{
+		x = b;
+		y = a;
+	}
+
+	// Několikanásobné použití out
+	public void GetValues(out Int32 x, out Int32 y)
+	{
+		Console.WriteLine("Enter the first value: ");
+		x = Convert.ToInt32(Console.ReadLine());
+
+		Console.WriteLine("Enter the second value: ");
+		y = Convert.ToInt32(Console.ReadLine());
+	}
+
+	public Int32 Sum(in Int32 a, in Int32 b)
+	{
+		//a++; // Nelze
+		return a + b;
+	}
+
+	// Lokálně definované funkce, tj. funkce definované uvnitř dalších funkcí.
+	public void PrintNames(List<string> childrens)
+	{
+		Int32 i = 1; // Globální proměnná pro PrintNames metodu.
+		foreach (var children in childrens)
 		{
-			return num1 > num2 ? num1 : num2;
+			PrintName(children); // Volání lokální funkce.
 		}
 
-		// Rekurzivní funkce
-		public Int32 Factorial(Int32 num)
+		void PrintName(string name)
 		{
-			// 0 a 1.
-			if (num < 2)
-			{
-				return 1;
-			}
-
-			return Factorial(num - 1) * num;
-			// Hrozná prasárna pro zásobník volání.
+			Console.WriteLine($"Name is {name} and he/she is my {i}. children");
+			i++; // Inkrementujeme globální proměnnou z volající funkce.
 		}
+	}
 
-		// Předávání parametrů hodnotou
-		// Dochází dvakrát k boxingu (zabalení), konkrétně
-		// hodnoty x a y.
-		public void Swap(Int32 x, Int32 y)
-		{
-			var temp = x;
-			x = y;    /* put y into x */
-			y = temp; /* put temp into y */
-			// return; // nepovinné a je odebráno při automatickém formátování kódu.
-		}
+	public void SomeOtherFunction()
+	{
+		PrintNames(new List<string>() { "Jméno", "Jméno2" });
+		//PrintName();  // Zde nedostupné.
+	}
 
-		// Přetížená funkce prohození s předáním referencí
-		public void Swap(ref Int32 x, ref Int32 y)
-		{
-			var temp = x;
-			x = y;    /* put y into x */
-			y = temp; /* put temp into y */
-		}
-
-		// Parametry definované s ref musí být při volání funkce
-		// definovány taktéž jako ref, tj. jako reference.
-		public void SwapRef(ref Int32 x, ref Int32 y)
-		{
-			var temp = x;
-			x = y;    /* put y into x */
-			y = temp; /* put temp into y */
-		}
-
-		// Několika násobný výstup z funkce
-		// Klíčové slovo out říká kompilátoru,
-		// že proměnná bude inicializována později.
-		public Int32 GetSomeValue(out Int32 x)
-		{
-			x = 5;
-			return x;
-		}
-
-		// Definice swap funkce (prohození) pomocí ref a out parametrů.
-		public void SwapOut(ref Int32 a, ref Int32 b, out Int32 x, out Int32 y)
-		{
-			var temp = a;
-			x = b;
-			y = temp;
-		}
-
-		// Několikanásobné použití out
-		public void GetValues(out Int32 x, out Int32 y)
-		{
-			Console.WriteLine("Enter the first value: ");
-			x = Convert.ToInt32(Console.ReadLine());
-
-			Console.WriteLine("Enter the second value: ");
-			y = Convert.ToInt32(Console.ReadLine());
-		}
-
-		public Int32 Sum(in Int32 a, in Int32 b)
-		{
-			//a++; // Nelze
-			return a + b;
-		}
-
-		// Implicitně definované funkce, tj. funkce definované uvnitř dalších funkcí.
-		public void PrintNames(List<string> childrens)
-		{
-			Int32 i = 1; // Globální proměnná pro PrintNames metodu.
-			foreach (var children in childrens)
-			{
-				PrintName(children); // Volání lokální funkce.
-			}
-
-			void PrintName(string name)
-			{
-				Console.WriteLine($"Name is {name} and he/she is my {i}. children");
-				i++; // Inkrementujeme globální proměnnou z volající funkce.
-			}
-		}
-
-		public void SomeOtherFunction()
-		{
-			PrintNames(new List<string>() { "Jméno", "Jméno2" });
-			//PrintName();  // Zde nedostupné.
-		}
-
-		// Výchozí hodnoty parametrů
-		// U vstupních parametrů funkcí můžeme definovat výchozí hodnotu.
-		// Ta bude použita pokud u volání funkce nezadáme žádnou hodnotu,
-		// např. PrintNumber(3.14159) povede na PrintNumber(3.14159,3).
-		// Parametry s výchozí hodnotou musí stát vždy nejvíc vpravo a 
-		// nesmí být mezi nimi žádný bez výchozí hodnoty.
-		public void PrintNumber(double number, byte numberOfDecimalPlaces = 3)
-		{
-			Console.WriteLine($"Number: {number:numberOfDecimalPlaces}");
-		}
+	// Výchozí hodnoty parametrů
+	// U vstupních parametrů funkcí můžeme definovat výchozí hodnotu.
+	// Ta bude použita pokud u volání funkce nezadáme žádnou hodnotu,
+	// např. PrintNumber(3.14159) povede na PrintNumber(3.14159,3).
+	// Parametry s výchozí hodnotou musí stát vždy nejvíc vpravo a 
+	// nesmí být mezi nimi žádný bez výchozí hodnoty.
+	public void PrintNumber(double number, byte numberOfDecimalPlaces = 3)
+	{
+		Console.WriteLine($"Number: {number:numberOfDecimalPlaces}");
 	}
 }
